UIU-877: Add override possibility for single-item.
