+++ src/components/Loans/OpenLoans/components/Modals/Modals.js
@@ -5,8 +5,6 @@ import { ChangeDueDateDialog } from '@folio/stripes/smart-components';
 import { stripesShape } from '@folio/stripes/core';
 
 // eslint-disable-next-line import/no-extraneous-dependencies
-import BulkRenewalDialog from '@folio/users/src/components/BulkRenewalDialog';
-// eslint-disable-next-line import/no-extraneous-dependencies
 import PatronBlockModal from '@folio/users/src/components/PatronBlock/PatronBlockModal';
 
 class Modals extends React.Component {
@@ -15,17 +13,10 @@ class Modals extends React.Component {
     loans: PropTypes.arrayOf(PropTypes.object).isRequired,
     activeLoan: PropTypes.string,
     user: PropTypes.object.isRequired,
-    loanPolicies: PropTypes.object.isRequired,
-    errorMsg: PropTypes.object.isRequired,
     patronGroup: PropTypes.object.isRequired,
     checkedLoans: PropTypes.object.isRequired,
-    requestCounts: PropTypes.object.isRequired,
     patronBlockedModal: PropTypes.bool.isRequired,
-    bulkRenewalDialogOpen: PropTypes.bool.isRequired,
     patronBlocks: PropTypes.arrayOf(PropTypes.object).isRequired,
-    renewSuccess: PropTypes.arrayOf(PropTypes.object).isRequired,
-    renewFailure: PropTypes.arrayOf(PropTypes.object).isRequired,
-    hideBulkRenewalDialog: PropTypes.func.isRequired,
     hideChangeDueDateDialog: PropTypes.func.isRequired,
     changeDueDateDialogOpen: PropTypes.bool.isRequired,
     onClosePatronBlockedModal: PropTypes.func.isRequired,
@@ -40,7 +31,6 @@ class Modals extends React.Component {
 
     const { stripes } = props;
     this.connectedChangeDueDateDialog = stripes.connect(ChangeDueDateDialog);
-    this.connectedBulkRenewalDialog = stripes.connect(BulkRenewalDialog);
   }
 
   render() {
@@ -48,17 +38,10 @@ class Modals extends React.Component {
       stripes,
       loans,
       user,
-      loanPolicies,
-      errorMsg,
-      requestCounts,
-      renewSuccess,
-      renewFailure,
-      bulkRenewalDialogOpen,
       changeDueDateDialogOpen,
       activeLoan,
       checkedLoans,
       hideChangeDueDateDialog,
-      hideBulkRenewalDialog,
       patronBlocks,
       patronBlockedModal,
       onClosePatronBlockedModal,
@@ -72,17 +55,6 @@ class Modals extends React.Component {
 
     return (
       <React.Fragment>
-        <this.connectedBulkRenewalDialog
-          user={user}
-          stripes={stripes}
-          errorMessages={errorMsg}
-          loanPolicies={loanPolicies}
-          open={bulkRenewalDialogOpen}
-          failedRenewals={renewFailure}
-          requestCounts={requestCounts}
-          successRenewals={renewSuccess}
-          onClose={hideBulkRenewalDialog}
-        />
         <this.connectedChangeDueDateDialog
           user={user}
           stripes={stripes}
